% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_simulated_dataset.R
\name{create_simulated_dataset}
\alias{create_simulated_dataset}
\title{Simulate a transcript-level single-cell counts table}
\usage{
create_simulated_dataset(
  dataset_name,
  number_of_genes = 2000,
  cells_in_groups = c(200, 200),
  isoform_number_distribution = data.frame(isoforms_in_gene = 1:10, proportion = rep(0.1,
    10)),
  gamma_shape = 1.5,
  gamma_scale = 0.5,
  proportion_of_genes_with_DTU = 0.3,
  isoforms_to_swap = c(1, 2),
  outlier_gene_proportion = 0,
  gene_outliers_lognormal_location = 2.5,
  gene_outliers_lognormal_scale = 0.3,
  proportion_of_genes_with_DGE = 0,
  DGE_lognormal_location = 0.8,
  DGE_lognormal_scale = 0.8,
  counts_distribution_type = "negative_binomial",
  negative_binomial_low_bound = 0.2,
  negative_binomial_high_bound = 0.7,
  proportion_of_cells_with_inflated_counts = 0,
  cell_outliers_lognormal_location = 2.3,
  cell_outliers_lognormal_scale = 0.4,
  save_filepath = NA,
  skip_distribution_dataframe_check = F
)
}
\arguments{
\item{dataset_name}{The name of the dataset generated.}

\item{number_of_genes}{The total number of genes to generate.}

\item{cells_in_groups}{A vector defining the number of cell groups, as well as the number of cells in each group.
For example, c(100, 200) would create two cell groups of 100 and 200 cells respectively.}

\item{isoform_number_distribution}{A dataframe with two columns. The 'isoforms_in_gene" column contains the number of isoforms per gene,
and the 'proportion' column contains a float between 0-1 that determines the proportion of genes to have that isoform number.}

\item{gamma_shape}{The shape parameter of the gamma distribution to pull gene means from.}

\item{gamma_scale}{The scale parameter of the gamma distribution to pull gene means from.}

\item{proportion_of_genes_with_DTU}{A decimal between 0 and 1 that defines the proportion of all genes that should exhibit DTU.}

\item{isoforms_to_swap}{A two-element vector that defines which isoforms should have their proportions swapped to simulate DTU.
Isoforms are numbered by abundance, where isoform 1 contains the highest proportion of gene counts.}

\item{outlier_gene_proportion}{The proportion of all genes that should exhibit outlier counts (between 0 and 1).}

\item{gene_outliers_lognormal_location}{The location parameter for the log-normal curve to pull the gene outlier counts multiplier from.}

\item{gene_outliers_lognormal_scale}{The scale parameter for the log-normal curve to pull the gene outlier counts multiplier from.}

\item{proportion_of_genes_with_DGE}{The proportion of all genes that should exhibit DGE (between 0 and 1).}

\item{DGE_lognormal_location}{The location parameter for the log-normal curve to pull the DGE count multiplier from.}

\item{DGE_lognormal_scale}{The scale parameter for the log-normal curve to pull the DGE count multiplier from.}

\item{counts_distribution_type}{The curve to generate counts from. Either 'poisson' or 'negative_binomial'.}

\item{negative_binomial_low_bound}{The low bound for the size parameter for the negative binomial curve to generate counts from. Defaults to 0.2.}

\item{negative_binomial_high_bound}{The high bound for the size parameter for the negative binomial curve to generate counts from. Defaults to 0.7.}

\item{proportion_of_cells_with_inflated_counts}{The proportion of all cells that should exhibit outlier counts (between 0 and 1).}

\item{cell_outliers_lognormal_location}{The location parameter for the log-normal curve to pull the cell outlier counts multiplier from.}

\item{cell_outliers_lognormal_scale}{The scale parameter for the log-normal curve to pull the cell outlier counts multiplier from.}

\item{save_filepath}{If set, saves the simulated data as an RDS object at the filepath provided. Defaults to NA.}

\item{skip_distribution_dataframe_check}{Skips error checking of the dataframe provided in the isoform_number_distribution dataframe. Defaults to FALSE.}
}
\value{
A list object containing all simulation data and parameters.
}
\description{
This function wraps all other simulation functions to make a user-friendly one-liner that
creates a transcript-level single-cell dataset, possibly exhibiting DTU between cell groups.
}
