% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_prop_tables.R
\name{generate_prop_tables}
\alias{generate_prop_tables}
\title{Generate a list of transcript proportion tables per group}
\usage{
generate_prop_tables(
  transcript_counts_table,
  cell_labels_table,
  transcript_id_colname = "transcript_id",
  gene_id_colname = "gene_id",
  cell_labels_colname = "group"
)
}
\arguments{
\item{transcript_counts_table}{A dataframe with a column containing transcript IDs, a column with gene IDs for
each transcript, and counts for all transcripts in all cells, with cell IDs as subsequent column names.}

\item{cell_labels_table}{A dataframe with two columns: one listing all cell IDs, and the other listing the group
that the cell ID belongs to.}

\item{transcript_id_colname}{A string corresponding to the column name in transcript_counts_table where the transcript IDs are stored. Defaults to 'transcript_id'.}

\item{gene_id_colname}{A string corresponding to the column name in transcript_counts_table where the gene IDs are stored. Defaults to 'gene_id'.}

\item{cell_labels_colname}{A string corresponding to the column name in cell_labels_table where the group information is stored. Defaults to 'group'.}
}
\value{
A list of dataframes, with each dataframe corresponding to a specific group listed in cell_labels_table.
}
\description{
This function generates a list of dataframes that include information regarding each transcript's counts,
gene-level counts, proportion within all cells, group-specific proportions, and proportion differences.
}
