% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{make_plots}
\alias{make_plots}
\title{Make and save DIU plots}
\usage{
make_plots(
  proportions_tables,
  pvalue_object,
  transcript_id_colname,
  save_plots = NA,
  save_tables = NA
)
}
\arguments{
\item{proportions_tables}{The list of dataframes created using generate_prop_tables().}

\item{pvalue_object}{The p-value object created from get_permutation_pvals().}

\item{transcript_id_colname}{A string corresponding to the column name in transcript_counts_table where the transcript IDs are stored.}

\item{save_plots}{Default NA. If specified, the path to the directory that plots should be saved in.}

\item{save_tables}{Default NA. If specified, the path to the directory that the tables used to make the plots
should be saved in. Useful if you'd like to recreate the plots yourself or in another program.}
}
\value{
A list containing two further lists. The first is a list of ggplot objects, each corresponding to a plot for a cell grouping.
The second contains the
}
\description{
Creates a differential isoform usage plot for each group.
}
