% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots_proportions.R
\name{make_plots_proportion}
\alias{make_plots_proportion}
\title{Makes proportion plots for a specified gene}
\usage{
make_plots_proportion(
  prop_table,
  gene_of_interest,
  output_folder_gene,
  plot_width = 12,
  plot_height = 12,
  plot_dpi = 600
)
}
\arguments{
\item{prop_table}{The output of \code{generate_prop_tables}.}

\item{gene_of_interest}{Transcripts belonging to this gene will be plotted.}

\item{output_folder_gene}{A path to the folder to save the plots to.}

\item{plot_width, plot_height, plot_dpi}{Plotting arguments supplied to \code{ggplot2::ggsave()}. Plots by default
are saved with a width and height of 12, at 600 DPI.}
}
\value{
A list containing a ggplot2 plot objects.
}
\description{
Makes a set of proportion plots for transcripts within a specified gene.
For use within a plotting wrapper function.
}
