% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{make_plots}
\alias{make_plots}
\title{Make and save DIU plots}
\usage{
make_plots(
  transcript_counts_table,
  cell_labels_table,
  transcript_id_colname,
  gene_id_colname,
  cell_labels_colname,
  pvalue_object,
  gene_of_interest = NA,
  pval_cutoff = 0.05,
  generate_UMAPs = TRUE,
  save_to_folder = NA
)
}
\arguments{
\item{transcript_counts_table}{A dataframe with a column containing transcript IDs, a column with gene IDs for
each transcript, and counts for all transcripts in all cells, with cell IDs as subsequent column names.}

\item{cell_labels_table}{A dataframe with two columns: one listing all cell IDs, and the other listing the group
that the cell ID belongs to.}

\item{transcript_id_colname}{A string corresponding to the column name in transcript_counts_table where the
transcript IDs are stored.}

\item{gene_id_colname}{A string corresponding to the column name in transcript_counts_table where the gene IDs are
stored. Defaults to 'gene_id'.}

\item{cell_labels_colname}{A string corresponding to the column name in cell_labels_table where the group information
is stored. Defaults to 'group'.}

\item{pvalue_object}{The p-value object created from \code{get_permutation_pvals}.}

\item{gene_of_interest}{The name of a gene of interest to create additional plots for. If set, proportion UMAPs
and transcript proportion plots are created for the gene. Defaults to NA.}

\item{pval_cutoff}{The p-value threshold for determining significance. Defaults to 0.05}

\item{generate_UMAPs}{A bool determining whether UMAPs should be calculated/plotted. UMAP generation can take
a while for larger datasets, so the option is provided to skip this step if computational resources are limited.
Default TRUE.}

\item{save_to_folder}{A string corresponding to the filepath to the folder that all outputs will be saved to.
Defaults to NA.}
}
\value{
A list containing the plots generated, as well as the data used to create these plots.
}
\description{
A wrapper function around multiple plotting functions.
First runs \code{generate_prop_tables}, then creates DTU plots, proportion plots, and UMAPs.
}
